Add Lowering Functions for Tag Tests, Payloads, and Iterators
- Introduced new lowering functions in `lower_expr_misc.go` to handle tag test expressions, tag payload extractions, iterator initializations, and iterator next expressions.
- Moved existing lowering functions from `lower_expr.go` to the new file for better organization and separation of concerns.
- Enhanced the lowering process to support await and spawn expressions, improving the overall functionality of the MIR lowering phase.
- Updated error handling to ensure proper diagnostics for unexpected payload types during lowering operations.
